requirements/req-dev.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.1.1
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==5.7.1
pandas==1.0.1